+++ pyincore/analyses/joplinempiricalbuildingrestoration/__init__.py
@@ -5,5 +5,5 @@
 # and is available at https://www.mozilla.org/en-US/MPL/2.0/
 
 
-from pyincore.analyses.joplinempiricalrestoration.joplinempiricalrestoration import JoplinEmpiricalRestoration
-from pyincore.analyses.joplinempiricalrestoration.joplinempirrestor_util import JoplinEmpirRestorUtil
+from pyincore.analyses.joplinempiricalbuildingrestoration.joplinempiricalbuildingrestoration import JoplinEmpiricalBuildingRestoration
+from pyincore.analyses.joplinempiricalbuildingrestoration.joplinempirrestor_util import JoplinEmpirRestorUtil
